improve performance
For every dash in the component, a new style was created and thus
and a new object is allocated in memory which also has to be sent
over the native bridge. This is a performance intensive task
and may lead to jank.

See https://facebook.github.io/react-native/docs/stylesheet.html
for further information.

This commit introduces the use of StyleSheet.create to create
reusable styles and stores them by an id to reference them later
for dashes with the same styling.